# backend/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Liveprofit API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analysis-tasks": {
      "post": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_analysis_tasks_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SingleStockCreateRequest"
                  },
                  {
                    "$ref": "#/components/schemas/MarketWideCreateRequest"
                  }
                ],
                "discriminator": {
                  "propertyName": "task_type",
                  "mapping": {
                    "SINGLE_STOCK": "#/components/schemas/SingleStockCreateRequest",
                    "MARKET_WIDE": "#/components/schemas/MarketWideCreateRequest"
                  }
                },
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TaskCreatedData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_analysis_tasks_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TaskListData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-tasks/{task_id}": {
      "delete": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "Delete Task",
        "description": "删除任务记录（仅终态任务可删；FK 级联删除报告与 Outbox，DELETE 返回 200 envelope）。",
        "operationId": "delete_task_api_v1_analysis_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DeleteResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_analysis_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TaskDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "Cancel Task",
        "operationId": "cancel_task_api_v1_analysis_tasks__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TaskDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-tasks/{task_id}/rerun": {
      "post": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "Rerun Task",
        "description": "终态任务从指定节点重跑（单Agent重跑与提示词编辑方案 3.4）。\n\n路由层前置校验（FS/拓扑，服务保持 DB 纯净）：节点在任务拓扑内（404）、\nscreening 任务个股层 v1 限制（409）、entry checkpoint 存在（409）。",
        "operationId": "rerun_task_api_v1_analysis_tasks__task_id__rerun_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RerunTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TaskDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-tasks/{task_id}/execution-logs": {
      "get": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "Get Execution Logs",
        "operationId": "get_execution_logs_api_v1_analysis_tasks__task_id__execution_logs_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ExecutionLogsDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analysis-tasks/{task_id}/graph-topology": {
      "get": {
        "tags": [
          "analysis-tasks"
        ],
        "summary": "Get Graph Topology",
        "operationId": "get_graph_topology_api_v1_analysis_tasks__task_id__graph_topology_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_GraphTopologyDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analysis-tasks/{task_id}/events": {
      "get": {
        "tags": [
          "analysis-events"
        ],
        "summary": "Task Events",
        "operationId": "task_events_api_v1_analysis_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SSEContract"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis-tasks/{task_id}/report": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_v1_analysis_tasks__task_id__report_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ReportDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisOptions": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "AnalysisOptions",
        "description": "接口定义的其余选项（扩展点，服务层复验）。"
      },
...
      "Envelope_DeleteResultData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteResultData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[DeleteResultData]"
      },
...
      "Envelope_ExecutionLogsDTO_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExecutionLogsDTO"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ExecutionLogsDTO]"
      },
      "Envelope_GraphTopologyDTO_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GraphTopologyDTO"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[GraphTopologyDTO]"
      },
...
      "Envelope_ReportDTO_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReportDTO"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ReportDTO]"
      },
...
      "Envelope_TaskCreatedData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskCreatedData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[TaskCreatedData]"
      },
      "Envelope_TaskDTO_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskDTO"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[TaskDTO]"
      },
      "Envelope_TaskListData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskListData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[TaskListData]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketWideCreateRequest": {
        "properties": {
          "task_type": {
            "type": "string",
            "const": "MARKET_WIDE",
            "title": "Task Type"
          },
          "ticker": {
            "type": "null",
            "title": "Ticker"
          },
          "requested_trade_date": {
            "type": "string",
            "format": "date",
            "title": "Requested Trade Date"
          },
          "selected_layers": {
            "items": {
              "type": "string",
              "enum": [
                "market",
                "sector",
                "stock",
                "screening",
                "position"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Selected Layers"
          },
          "analysis_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "strategy_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Version Id"
          },
          "portfolio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Id"
          },
          "expected_portfolio_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Portfolio Version"
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "requested_trade_date",
          "selected_layers"
        ],
        "title": "MarketWideCreateRequest"
      },
...
      "RerunTaskRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id",
            "description": "重跑起点节点 id，如 \"market:CN News Analyst\""
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "RerunTaskRequest"
      },
...
      "SSEContract": {
        "properties": {
          "business": {
            "$ref": "#/components/schemas/BusinessEventData"
          },
          "reset": {
            "$ref": "#/components/schemas/ResetEventData"
          },
          "heartbeat": {
            "$ref": "#/components/schemas/HeartbeatEventData"
          }
        },
        "type": "object",
        "required": [
          "business",
          "reset",
          "heartbeat"
        ],
        "title": "SSEContract",
        "description": "SSE 协议帧集合（OpenAPI 文档用途：运行时为 text/event-stream 分帧）。"
      },
      "SingleStockCreateRequest": {
        "properties": {
          "task_type": {
            "type": "string",
            "const": "SINGLE_STOCK",
            "title": "Task Type"
          },
          "ticker": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Ticker",
            "description": "单股分析目标"
          },
          "requested_trade_date": {
            "type": "string",
            "format": "date",
            "title": "Requested Trade Date"
          },
          "selected_layers": {
            "items": {
              "type": "string",
              "enum": [
                "market",
                "sector",
                "stock",
                "screening",
                "position"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Selected Layers"
          },
          "analysis_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "ticker",
          "requested_trade_date",
          "selected_layers"
        ],
        "title": "SingleStockCreateRequest"
      },
...
      "TaskCreatedData": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "QUEUED",
              "RUNNING",
              "RETRYING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "CANCEL_REQUESTED"
            ],
            "title": "Status"
          },
          "requested_trade_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Trade Date"
          },
          "effective_trade_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Trade Date"
          },
          "date_correction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Correction"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url",
            "description": "服务端派生的 SSE 相对 URL"
          },
          "report_url": {
            "type": "string",
            "title": "Report Url",
            "description": "服务端派生的报告相对 URL"
          },
          "execution_logs_url": {
            "type": "string",
            "title": "Execution Logs Url",
            "description": "服务端派生的执行调用日志相对 URL"
          },
          "graph_topology_url": {
            "type": "string",
            "title": "Graph Topology Url",
            "description": "服务端派生的图拓扑相对 URL"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status",
          "requested_trade_date",
          "effective_trade_date",
          "date_correction",
          "events_url",
          "report_url",
          "execution_logs_url",
          "graph_topology_url"
        ],
        "title": "TaskCreatedData"
      },
...
      "TaskListData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskListItemDTO"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TaskListData"
      },
...
    }
  }
}